refactor: change the name of the file for the details screen
