+++ app/routes/home.tsx
@@ -1,6 +1,6 @@
 import type { Route } from "./+types/home";
 import { Welcome } from "../welcome/welcome";
-import RoadMapDetailsScreen from "~/roadmap/roadmap_details/roadmap_details";
+import RoadMapDetailsScreen from "~/roadmap/roadmap_details/roadmap_details_screen";
 
 export function meta({}: Route.MetaArgs) {
   return [
